requirements.txt
numpy==1.23.5
opencv_python==4.7.0.72
pandas==2.0.1
scikit_learn==1.2.2
tensorflow_cpu==2.12.0
tensorflow_hub==0.13.0
tqdm==4.65.0
